+++ Assets/PatchKit Patcher/Scripts/Background.cs
@@ -124,7 +124,11 @@ public class Background : MonoBehaviour
         _logger.LogDebug("On patcher data update.");
         var bannerData = data.BannerData;
 
-        if (IsNewBannerAvailable(data))
+        if (IsLocalBannerMissing(data))
+        {
+            AquireRemoteBanner(data);
+        }
+        else if (IsNewBannerAvailable(data))
         {
             AquireRemoteBanner(data);
         }
@@ -154,6 +158,11 @@ public class Background : MonoBehaviour
         MainAnimator.SetTrigger(AnimationSwitchTrigger);
     }
 
+    private bool IsLocalBannerMissing(Data data)
+    {
+        return !string.IsNullOrEmpty(CachedBannerModificationDate) && !File.Exists(CachedBannerPath);
+    }
+
     private bool IsNewBannerAvailable(Data data)
     {
         return !string.IsNullOrEmpty(data.BannerData.ImageUrl)
@@ -175,9 +184,7 @@ public class Background : MonoBehaviour
 
     private bool IsCachedBannerSameAsRemote(PatcherBannerData bannerData)
     {
-        var cachedModificationDate = CachedBannerModificationDate;
-
-        return bannerData.ModificationDate == cachedModificationDate;
+        return bannerData.ModificationDate == CachedBannerModificationDate;
     }
 
     private void ClearCachedBanner()