+++ CHANGELOG.md
@@ -31,6 +31,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Patcher will timeout if the downloading stopped sooner than after 5 minutes
 
+### Fixed
+- An edge case which caused the download to never terminate
+- Patcher wouldn't quit upon starting the Launcher
+- Patcher wouldn't re-download a manually removed banner file
+
 ## [3.10.1]
 ### Fixed
 - Updated torrent-client to fix the issue with paths with spaces in them