+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [3.10.2]
 ### Fixed
 - An edge case which caused the download to never terminate
 - Patcher wouldn't quit upon starting the Launcher